requirements.txt
cryptography
pyOpenSSL < 20
requests
signxml
xmlschema
